+++ docs/src/test.sh
@@ -35,7 +35,7 @@ for filepath in ${test_files}; do
     PYTHONPATH=${test_dir}:${PYTHONPATH}
 
     echo Testing ${filename}..
-    coverage run -p $(which uvicorn) chunli:app >${uvicorn_output_file} 2>&1 &
+    coverage run -p $(which gunicorn) chunli:app -k uvicorn.workers.UvicornWorker -c gunicorn_conf.py >${uvicorn_output_file} 2>&1 &
     coverage run -p $(which uvicorn) --port 8001 index_hello_app:app \
         >${uvicorn_hello_output_file} 2>&1 &
     sleep 5
@@ -70,7 +70,7 @@ for filepath in ${test_files}; do
     $md5_cmd ${output_file2} ${output_tmpfile2} > ${checksum_file2}
 
     ps ax | (ps ax | awk "/uvicorn index_hello_app:app/ {print \$1}" | xargs kill -SIGTERM 2>/dev/null)
-    ps ax | (ps ax | awk "/uvicorn chunli:app/ {print \$1}" | xargs kill -SIGTERM 2>/dev/null)
+    ps ax | (ps ax | awk "/gunicorn chunli:app/ {print \$1}" | xargs kill -SIGTERM 2>/dev/null)
 
     output=$(sed -r -e 's/(.*) .*/\1/g' ${checksum_file} | uniq | wc -l)